moved product-details from components to pages, cleaned up css, and removed all commented out code #2
