+++ server/public/styles.css
@@ -41,8 +41,8 @@ body {
 
 .accordion-button {
   padding: 0;
-  font-size: calc(0.5rem + 1.25vw);
   font-weight: bold;
+  font-size: 1.25rem;
 }
 
 .accordion-body {
@@ -71,10 +71,6 @@ body {
   top: 0;
 }
 
-.accordion-button:focus {
-  border-color: white;
-}
-
 .p-name,
 .p-price {
   font-size: calc(0.5rem + 1vw);
@@ -84,10 +80,6 @@ a:hover {
   text-shadow: 0.5px 0.5px 1px rgb(183, 190, 193);
   color: black;
 }
-
-/* .list-group {
-  flex: 1 1 0;
-} */
 
 .p-img {
   max-height: 95%;
